refactor: move component into component folder with new css file
